requirements.txt
asgiref==3.4.1
boto3==1.16.6
botocore==1.19.63
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.6
django-allauth==0.44.0
django-ckeditor==6.1.0
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-extensions==3.1.1
django-js-asset==1.2.2
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-extra-fields==3.0.2
django-filter==2.4.0
django-rest-passwordreset==1.2.0
django-storages==1.10.1
djangorestframework==3.12.4
drf-yasg==1.20.0
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==21.0
Pillow==8.2.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
qrcode==6.1
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
s3transfer==0.3.7
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
whitenoise==5.2.0
dj-stripe==2.5.1
